requirements.txt
ansicon==1.89.0
arrow==1.2.3
asgiref==3.5.2
blessed==1.19.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cryptography==38.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.3
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-filter==22.1
django-parler==2.3
django-picklefield==3.1
django-q==1.3.9
django-rosetta==0.9.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
environs==9.5.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jinxed==1.2.0
MarkupSafe==2.1.1
marshmallow==3.19.0
oauthlib==3.2.2
packaging==22.0
Pillow==9.3.0
polib==1.1.1
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
redis==3.5.3
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0

